requirements.txt
annotated-types==0.5.0
anyio==3.7.1
azure-core==1.30.0
azure-data-tables==12.4.4
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
isodate==0.6.1
limits==3.8.0
multidict==6.0.5
openai==1.11.1
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
requests==2.31.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
wrapt==1.16.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zipp==3.15.0
